requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
sqlalchemy==2.0.44
pydantic==2.12.5
pymongo==4.15.5
motor==3.7.1
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
cryptography==46.0.3
email-validator==2.3.0
python-multipart==0.0.20
fastapi-mail==1.6.1

